fix/feat: SCD overhaul, CORS patch, admin roles endpoint, and operational period group routing
SCD versioning:
- Migrate version_from/version_to from LocalDate to LocalDateTime across
  Person, Vehicle, Route and ScheduleTemplate entities and services;
  capture a single LocalDateTime.now() for atomic version transitions
- Remove LocalDate.now().minusDays(1) antipattern that caused
  versionTo < versionFrom on same-day operations
- Add *Group SCD anchor entities (PersonGroup, RouteGroup, VehicleGroup,
  ScheduleTemplateGroup) and their repositories
- Add Flyway V15 (DATE→TIMESTAMPTZ migration) and V16 (bad-data cleanup)

Bug fixes:
- Add PATCH to CORS allowedMethods in SecurityConfig to fix 403 on
  /reactivate endpoints
- Fix null active on user creation: add this.active = true to User
  custom constructor (@Builder.Default does not apply to explicit ctors)
- Fix AdminUserService compilation error: LocalDate → LocalDateTime import
- Fix GET /api/routes/{groupId}/operational-periods 404: controller path
  variable changed from Integer routeId to Long groupId; service now
  resolves current SCD version via routeService.findByGroupId()

New features:
- Add GET /api/admin/roles endpoint (AdminRoleController + RoleDTO) that
  returns assignable roles excluding SUPER_ADMIN
- Add AppRoleRepository.findAssignable() query
- POST /api/routeOperation/all now returns 200 with routes.initialized.none
  when count=0 instead of a misleading 201

Removals:
- Delete ScheduleTemplateVersion entity, service, controller, repository,
  DTOs, requests and all associated tests (feature consolidated elsewhere)

Tests:
- Update RouteOperationalPeriodControllerTest and ServiceTest to use
  Long groupId (10L) and routeService.findByGroupId()
- Update all affected service/controller tests to reflect SCD group model

+++ src/main/java/com/example/tgs_dev/controller/request/ScheduleTemplateRequest.java
@@ -5,9 +5,16 @@ import jakarta.validation.constraints.NotNull;
 
 import java.time.LocalTime;
 
+/**
+ * Request body for creating or updating a schedule template.
+ *
+ * <p>{@code routeId} and {@code secondaryRouteId} are
+ * {@link com.example.tgs_dev.entity.RouteGroup} ids (stable business identity),
+ * not the surrogate version ids.
+ */
 public record ScheduleTemplateRequest(
-        @NotNull Integer routeId,
-        Integer secondaryRouteId,
+        @NotNull Long   routeId,
+        Long            secondaryRouteId,
         @NotBlank String templateNumber,
         @NotBlank String name,
         @NotNull LocalTime startTime,
