+++ src/main/java/com/example/tgs_dev/controller/request/VehicleRequest.java
@@ -2,8 +2,14 @@ package com.example.tgs_dev.controller.request;
 
 import jakarta.validation.constraints.NotBlank;
 
+/**
+ * Request body for creating or updating a vehicle.
+ *
+ * <p>{@code ownerId} is the {@link com.example.tgs_dev.entity.PersonGroup} id
+ * (stable business identity), not the surrogate version id.
+ */
 public record VehicleRequest(
         @NotBlank String vehicleNumber,
         @NotBlank String licensePlate,
-        Integer ownerId
+        Long ownerId
 ) {}
